로그아웃 API, 가계부 내역 수정 API, 가계부 내역 삭제 API, 가계부 내역 삭제 취소 API 구현

+++ src/main/java/com/habin/payhere_task/house_hold/repository/HouseHoldRepository.java
@@ -3,5 +3,5 @@ package com.habin.payhere_task.house_hold.repository;
 import com.habin.payhere_task.house_hold.entity.HouseHold;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface HouseHoldRepository extends JpaRepository<HouseHold, Integer>, QHouseHoldRepository {
+public interface HouseHoldRepository extends JpaRepository<HouseHold, Long>, QHouseHoldRepository {
 }
